Change from dev-master to semver
As currently it forces people to use dev-master, and so enable `'minimum-stability': 'dev'` which is bad. If you do need to use a dev-master version to get a bugfix before it has been tagged, it is better to do the require with `"nikic/fast-route": "dev-master as 0.2.1",` as then that shouldn't clash with other people's dependencies.